% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{run_omics_evaluation}
\alias{run_omics_evaluation}
\title{Perform omics datasets evaluation}
\usage{
run_omics_evaluation(data_dir = NULL, x2 = NULL, sample_list = NULL,
  data_type = "protein", use_class = NULL, class_color = NULL,
  out_dir = "./", cpu = 0, missing_value_cutoff = 0.5)
}
\arguments{
\item{data_dir}{A folder contains multiple omics datasets in tsv format.
Row: genes/proteins, column: samples.}

\item{x2}{A tsv format file which contains protein or gene expression data.
When the value of parameter is not NULL, the data in this file will be used
to perform correlation analysis between this dataset and the dataset in
data_dir}

\item{sample_list}{A file in tsv format contains sample class, batch and
order information.}

\item{data_type}{The quantification data type in folder data_dir:
protein, gene. Default is protein.}

\item{use_class}{The class of samples which will be used to perform
correlation analysis}

\item{class_color}{The color for class.}

\item{out_dir}{Output folder}

\item{cpu}{The number of CPUs. Default is 0 and all available CPUs will be
used.}

\item{missing_value_cutoff}{The cutoff of missing value filtering. Default
is 0.5.}
}
\value{
The path of HTML report.
}
\description{
Perform omics datasets evaluation
}
\author{
Bo Wen \email{wenbostar@gmail.com}
}
